[androsdk-1541] Move from main package to androidTest
