+++ core/src/androidTest/java/org/hisp/dhis/android/realservertests/schema/tests/EnumUpdatesCheckerRealIntegrationShould.kt
@@ -25,12 +25,12 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.common
+package org.hisp.dhis.android.realservertests.schema.tests
 
 import org.hisp.dhis.android.core.BaseRealIntegrationTest
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.D2Factory
-import org.hisp.dhis.android.core.common.schema.Schema
+import org.hisp.dhis.android.core.common.*
 import org.hisp.dhis.android.core.data.server.RealServerMother
 import org.hisp.dhis.android.core.enrollment.EnrollmentStatus
 import org.hisp.dhis.android.core.event.EventStatus
@@ -41,8 +41,11 @@ import org.hisp.dhis.android.core.validation.MissingValueStrategy
 import org.hisp.dhis.android.core.validation.ValidationRuleImportance
 import org.hisp.dhis.android.core.validation.ValidationRuleOperator
 import org.hisp.dhis.android.core.visualization.*
+import org.hisp.dhis.android.realservertests.schema.Schema
+import org.hisp.dhis.android.realservertests.schema.SchemaCall
 import org.junit.Assert
 import org.junit.Before
+import org.junit.Test
 
 class EnumUpdatesCheckerRealIntegrationShould : BaseRealIntegrationTest() {
     private lateinit var d2: D2
@@ -53,10 +56,10 @@ class EnumUpdatesCheckerRealIntegrationShould : BaseRealIntegrationTest() {
         d2 = D2Factory.forNewDatabase()
     }
 
-    // @Test
+    @Test
     fun check_no_enum_have_been_updated_on_server() {
         d2.userModule().blockingLogIn(username, password, RealServerMother.url2_38)
-        val schemas: List<Schema> = getD2DIComponent(d2).schemaCall().download().blockingGet()
+        val schemas: List<Schema> = SchemaCall(d2.retrofit()).download().blockingGet()
         val constantsMap: Map<String, List<String>?> = schemas.flatMap { schema ->
             schema.properties.filter { it.propertyType == "CONSTANT" }
         }.toSet().associate { fullKlassToSimpleKlass(it.klass) to it.constants }
