+++ core/src/main/java/org/hisp/dhis/android/core/arch/d2/internal/D2DIComponent.java
@@ -50,7 +50,6 @@ import org.hisp.dhis.android.core.attribute.internal.AttributePackageDIModule;
 import org.hisp.dhis.android.core.category.CategoryOption;
 import org.hisp.dhis.android.core.category.internal.CategoryPackageDIModule;
 import org.hisp.dhis.android.core.common.internal.CommonPackageDIModule;
-import org.hisp.dhis.android.core.common.schema.SchemaCall;
 import org.hisp.dhis.android.core.configuration.internal.ConfigurationPackageDIModule;
 import org.hisp.dhis.android.core.configuration.internal.MultiUserDatabaseManagerForD2Manager;
 import org.hisp.dhis.android.core.constant.internal.ConstantPackageDIModule;
@@ -195,8 +194,6 @@ public interface D2DIComponent {
     PeriodHandler periodHandler();
     @VisibleForTesting
     InterpreterSelector interpreterSelector();
-    @VisibleForTesting
-    SchemaCall schemaCall();
 
     @Component.Builder
     interface Builder {
