+++ core/src/main/java/org/hisp/dhis/android/core/common/internal/CommonPackageDIModule.java
@@ -30,7 +30,6 @@ package org.hisp.dhis.android.core.common.internal;
 
 import org.hisp.dhis.android.core.arch.call.internal.GenericCallData;
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
-import org.hisp.dhis.android.core.common.schema.SchemaDIModule;
 import org.hisp.dhis.android.core.common.valuetype.devicerendering.internal.ValueTypeDeviceRenderingEntityDIModule;
 import org.hisp.dhis.android.core.common.valuetype.rendering.internal.ValueTypeRenderingEntityDIModule;
 import org.hisp.dhis.android.core.resource.internal.ResourceHandler;
@@ -42,7 +41,6 @@ import dagger.Reusable;
 import retrofit2.Retrofit;
 
 @Module(includes = {
-        SchemaDIModule.class,
         ValueTypeDeviceRenderingEntityDIModule.class,
         ValueTypeRenderingEntityDIModule.class
 })
